requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0

httpx==0.28.1
pydantic==2.12.4
python-dotenv==1.0.1

black==25.11.0
ruff==0.14.5
pytest==8.3.2
